// clients/js/src/generated/types/baseAssetLinkedLifecycleHook.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import { Option, OptionOrNullable, PublicKey } from '@metaplex-foundation/umi';
import {
  Serializer,
  array,
  option,
  publicKey as publicKeySerializer,
  struct,
} from '@metaplex-foundation/umi/serializers';
import {
  BaseExtraAccount,
  BaseExtraAccountArgs,
  BasePluginAuthority,
  BasePluginAuthorityArgs,
  ExternalPluginAdapterSchema,
  ExternalPluginAdapterSchemaArgs,
  getBaseExtraAccountSerializer,
  getBasePluginAuthoritySerializer,
  getExternalPluginAdapterSchemaSerializer,
} from '.';

export type BaseAssetLinkedLifecycleHook = {
  hookedProgram: PublicKey;
  extraAccounts: Option<Array<BaseExtraAccount>>;
  dataAuthority: Option<BasePluginAuthority>;
  schema: ExternalPluginAdapterSchema;
};

export type BaseAssetLinkedLifecycleHookArgs = {
  hookedProgram: PublicKey;
  extraAccounts: OptionOrNullable<Array<BaseExtraAccountArgs>>;
  dataAuthority: OptionOrNullable<BasePluginAuthorityArgs>;
  schema: ExternalPluginAdapterSchemaArgs;
};

export function getBaseAssetLinkedLifecycleHookSerializer(): Serializer<
  BaseAssetLinkedLifecycleHookArgs,
  BaseAssetLinkedLifecycleHook
> {
  return struct<BaseAssetLinkedLifecycleHook>(
    [
      ['hookedProgram', publicKeySerializer()],
      ['extraAccounts', option(array(getBaseExtraAccountSerializer()))],
      ['dataAuthority', option(getBasePluginAuthoritySerializer())],
      ['schema', getExternalPluginAdapterSchemaSerializer()],
    ],
    { description: 'BaseAssetLinkedLifecycleHook' }
  ) as Serializer<
    BaseAssetLinkedLifecycleHookArgs,
    BaseAssetLinkedLifecycleHook
  >;
}